// frontend/src/api/generated/models/ProductDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductDto = {
    id?: number;
    code?: string | null;
    name?: string | null;
    price?: number | null;
    createdAt?: string;
    createdBy?: string | null;
    updatedAt?: string | null;
    updatedBy?: string | null;
};